Refactor a Trainer object out of the classifier.py main program.  In addition to the model, this saves and loads the optimizer and the number of epochs trained.  Purpose: to make it so that it is easy to checkpoint model training the same way the charlm is checkpointed
