+++ stanza/models/classifiers/cnn_classifier.py
@@ -449,75 +449,6 @@ class CNNClassifier(nn.Module):
         return out
 
 
-# TODO: make some of the following methods part of the class
-
-# TODO: all this code is basically the same as for POS and NER.  Should refactor
-def save(filename, model, skip_modules=True):
-    save_dir = os.path.split(filename)[0]
-    os.makedirs(save_dir, exist_ok=True)
-    model_state = model.state_dict()
-    # skip saving modules like pretrained embeddings, because they are large and will be saved in a separate file
-    if skip_modules:
-        skipped = [k for k in model_state.keys() if k.split('.')[0] in model.unsaved_modules]
-        for k in skipped:
-            del model_state[k]
-    params = {
-        'model': model_state,
-        'config': model.config,
-        'labels': model.labels,
-        'extra_vocab': model.extra_vocab,
-    }
-    torch.save(params, filename, _use_new_zipfile_serialization=False)
-    logger.info("Model saved to {}".format(filename))
-
-def load(filename, pretrain, charmodel_forward, charmodel_backward, elmo_model, foundation_cache=None):
-    try:
-        checkpoint = torch.load(filename, lambda storage, loc: storage)
-    except BaseException:
-        logger.exception("Cannot load model from {}".format(filename))
-        raise
-    logger.debug("Loaded model {}".format(filename))
-
-    # TODO: should not be needed when all models have this value set
-    setattr(checkpoint['config'], 'use_elmo', getattr(checkpoint['config'], 'use_elmo', False))
-    setattr(checkpoint['config'], 'elmo_projection', getattr(checkpoint['config'], 'elmo_projection', False))
-    setattr(checkpoint['config'], 'char_lowercase', getattr(checkpoint['config'], 'char_lowercase', False))
-    setattr(checkpoint['config'], 'charlm_projection', getattr(checkpoint['config'], 'charlm_projection', None))
-    setattr(checkpoint['config'], 'bert_model', getattr(checkpoint['config'], 'bert_model', None))
-    setattr(checkpoint['config'], 'bilstm', getattr(checkpoint['config'], 'bilstm', False))
-    setattr(checkpoint['config'], 'bilstm_hidden_dim', getattr(checkpoint['config'], 'bilstm_hidden_dim', 0))
-    setattr(checkpoint['config'], 'maxpool_width', getattr(checkpoint['config'], 'maxpool_width', 1))
-
-    # TODO: the getattr is not needed when all models have this baked into the config
-    model_type = getattr(checkpoint['config'], 'model_type', 'CNNClassifier')
-
-    bert_model = checkpoint['config'].bert_model
-    bert_model, bert_tokenizer = load_bert(bert_model, foundation_cache)
-    if model_type == 'CNNClassifier':
-        extra_vocab = checkpoint.get('extra_vocab', None)
-        model = CNNClassifier(pretrain=pretrain,
-                              extra_vocab=extra_vocab,
-                              labels=checkpoint['labels'],
-                              charmodel_forward=charmodel_forward,
-                              charmodel_backward=charmodel_backward,
-                              elmo_model=elmo_model,
-                              bert_model=bert_model,
-                              bert_tokenizer=bert_tokenizer,
-                              args=checkpoint['config'])
-    else:
-        raise ValueError("Unknown model type {}".format(model_type))
-    model.load_state_dict(checkpoint['model'], strict=False)
-
-    logger.debug("-- MODEL CONFIG --")
-    for k in model.config.__dict__:
-        logger.debug("  --{}: {}".format(k, model.config.__dict__[k]))
-
-    logger.debug("-- MODEL LABELS --")
-    logger.debug("  {}".format(" ".join(model.labels)))
-
-    return model
-
-
 def label_text(model, text, batch_size=None):
     """
     Given a list of sentences, return the model's results on that text.
