+++ stanza/pipeline/sentiment_processor.py
@@ -9,7 +9,10 @@ TODO: a possible way to generalize this would be to make it a
 ClassifierProcessor and have "sentiment" be an option.
 """
 
+from types import SimpleNamespace
+
 import stanza.models.classifiers.cnn_classifier as cnn_classifier
+from stanza.models.classifiers.trainer import Trainer
 
 from stanza.models.common import doc
 from stanza.pipeline._constants import *
@@ -33,24 +36,23 @@ class SentimentProcessor(UDProcessor):
         charmodel_forward = pipeline.foundation_cache.load_charlm(forward_charlm_path)
         backward_charlm_path = config.get('backward_charlm_path', None)
         charmodel_backward = pipeline.foundation_cache.load_charlm(backward_charlm_path)
+        args = SimpleNamespace(cuda = use_gpu)
         # set up model
         # elmo does not have a convenient way to download intermediate
         # models the way stanza downloads charlms & pretrains or
         # transformers downloads bert etc
         # however, elmo in general is not as good as using a
         # transformer, so it is unlikely we will ever fix this
-        self._model = cnn_classifier.load(filename=config['model_path'],
-                                          pretrain=self._pretrain,
-                                          charmodel_forward=charmodel_forward,
-                                          charmodel_backward=charmodel_backward,
-                                          elmo_model=None,
-                                          foundation_cache=pipeline.foundation_cache)
+        trainer = Trainer.load(filename=config['model_path'],
+                               args=args,
+                               pretrain=self._pretrain,
+                               charmodel_forward=charmodel_forward,
+                               charmodel_backward=charmodel_backward,
+                               elmo_model=None,
+                               foundation_cache=pipeline.foundation_cache)
+        self._model = trainer.model
         # batch size counted as words
         self._batch_size = config.get('batch_size', SentimentProcessor.DEFAULT_BATCH_SIZE)
-
-        # TODO: move this call to load()
-        if use_gpu:
-            self._model.cuda()
 
     def process(self, document):
         sentences = document.sentences
